==> 3._train-test_lvl2_model_top12_recs.ipynb <==
import pandas as pd
import numpy as np
import gc

from rec_lib.utils import reduce_mem_usage, set_target, customers_prep, articles_prep, df_lvl2_prep, top_12_recs
from rec_lib.utils import get_features, get_preds_result, col_convert_int
from rec_lib.models import get_lvl2_model_preds, get_lvl2_LGBM_model_preds
from rec_lib.metrics import precision_at_k, ap_k, recall, recall_at_k

import warnings
warnings.filterwarnings("ignore")

%load_ext autoreload
%autoreload 2

### Загрузка данных

%%time
result = pd.read_parquet('archive/result.parquet')
customers = pd.read_parquet('archive/customers.parquet')
articles = pd.read_parquet('archive/articles.parquet')
transactions = pd.read_parquet('archive/transactions_train_for_power_bi.parquet')

### Подготовка данных

%%time
# преобразование датафрейма пользователей
customers = customers_prep(customers)

%%time
# преобразование датафрейма категорий
articles = articles_prep(articles)

### Собираем датафрейм для обучения модели 2го уровня

%%time
# добавим нужные признаки 
result = df_lvl2_prep(result)

%%time
# добавляем данные из датафремов покупателей и категорий
result = result.merge(customers, on='customer_id', how='left')
result = result.merge(articles, on='article_id', how='left')

del customers;
del articles;
gc.collect()

### Добавляем фичи
Добавляем следующие фичи:
1. сумма продаж по категориям товаров
2. средняя сумма покупок по категориям товаров
3. количество заказов покупателя с товарами категории
4. количество единиц товаров категории, приобретенное покупателем
5. сумма покупок покупателя по категории
6. средний чек покупателя по категории
7. общая сумма покупок покупателя
8. средний чек покупателя

%%time
result = get_features(result, transactions)

#### Выбираем фичи для модели

# все используемые фичи
features = list(result)
features.remove('target')
features.remove('customer_id')
features.remove('article_id')

# категориальные фичи
cat_feats = features.copy()
cat_feats.remove('age')
cat_feats.remove('customer_id_short')
cat_feats.remove('article_id_short')

# убираем из категариальных фичей новые численные
cat_feats.remove('articles_buy_sum')
cat_feats.remove('articles_buy_avg')

cat_feats.remove('customers_articles_num_sum')
cat_feats.remove('customers_articles_num_count')

cat_feats.remove('customers_articles_sum')
cat_feats.remove('customers_articles_avg')

cat_feats.remove('customers_spent_sum')
cat_feats.remove('customers_spent_avg')

### Построение моделей и получение топ12 рекомендаций
Построим следующие модели классификаторов:
- model 1 - покупатели до 24 лет. Молодые люди с невысокой покупательной способностью, несформировавшимися окончательно привычками, без сильного влияния детей.
- model 2 - покупатели 25-44 года. Платежеспособные люди с устоявшимися привычками и существенным влиянием детей на выбор товаров.
- model 3 - покупатели 45-64 года. Платежеспособные люди с устоявшимися привычками и уже без сильного влияния детей на выбор товаров.
- model 4 - покупатели старше 65 лет. Люди пенсионного возраста с устоявшимися привычками, но невысокой покупательной способностью.

### Предсказания моделей CatBoost и их оценка

%%time
customers_id, article_id, target, preds = get_lvl2_model_preds(result, features, cat_feats)

%%time
# построение датафрейма для получения оценок предсказаний модели
filename = 'result'
result_test = get_preds_result(customers_id, article_id, target, preds, filename)

print('Recall_own_rec_12 : ', result_test.apply(lambda row: recall(row['top_12_recs'], row['actual_article_id_short']), axis=1).mean())

print('MAP_own_rec_12 : ', result_test.apply(lambda row: ap_k(row['top_12_recs'], row['actual_article_id_short'], 12), axis=1).mean())

### Предсказания моделей LGBM и их оценка

# преобразуем тип данных поля age для модели LGBM
result['age'] = result.apply(lambda row: col_convert_int(row['age']), axis=1)

%%time
customers_id, article_id, target, preds = get_lvl2_LGBM_model_preds(result, features, cat_feats)

%%time
# построение датафрейма для получения оценок предсказаний модели
filename = 'result'
result_test = get_preds_result(customers_id, article_id, target, preds, filename)

print('Recall_own_rec_12 : ', result_test.apply(lambda row: recall(row['top_12_recs'], row['actual_article_id_short']), axis=1).mean())

print('MAP_own_rec_12 : ', result_test.apply(lambda row: ap_k(row['top_12_recs'], row['actual_article_id_short'], 12), axis=1).mean())
